Reorganized approval components for cleaner code
